1. 增加注释；2. 优化获取记录详情方法；3. 增加单元测试

+++ MinCloud/CurrentUser.swift
@@ -13,7 +13,7 @@ import Result
 @objc(BaaSCurrentUser)
 open class CurrentUser: User {
 
-    override init(Id: Int) {
+    override init(Id: Int64) {
         super.init(Id: Id)
     }
